src/requirements.txt
json
jinja